Fully qualify `puppet` command for console refresh
Without this commit, the method that refreshes the Puppet Enterprise
Console was calling out to the Puppet binary without fully qualifying
the command. This was resulting in a failure of the task as the command
was not returning properly.

This commit fully qualifies the usage of the `puppet` binary
(/opt/puppetlabs/bin/puppet) to ensure that the proper binary is used
and will return successfully.